backend/requirements.txt
fastapi
uvicorn
pydantic
asyncpg  # Driver assíncrono para PostgreSQL
SQLAlchemy  # ORM (Opcional, mas recomendado para facilitar a interação com o banco de dados)
psycopg2-binary  # Driver síncrono para SQLAlchemy (alternativa a asyncpg com SQLAlchemy)
python-jose[cryptography]  # Para geração de tokens JWT
passlib[bcrypt]  # Para hashing de senhas